Enforce mandatory build and linting checks in CI while allowing relaxed local development
- Added FSLINT_WERROR option to CMakeLists.txt (default OFF) to treat warnings as errors.
- Added check-format target to CMakeLists.txt for dry-run formatting verification.
- Updated .clang-tidy to keep misc-include-cleaner and disable noisy/unavoidable binary parsing rules.
- Updated .github/workflows/build.yml to enforce WERROR, clang-tidy (with --warnings-as-errors='*'), and formatting in the pipeline.
- Fixed MSVC and clang-tidy warnings in the source code (uninitialized variables, narrowing conversions, constant conditional expressions).
- Added support for .gitkeep in DirectoryChecker to allow mandatory emptiness tests while tracking directories in Git.

+++ lib/zipper.cpp
@@ -78,11 +78,11 @@ std::vector<ZipFileEntry> Zipper::getEntries() const
 
         ZipFileEntry entry;
         entry.filename = filename.data();
-        entry.compression_method = file_info.compression_method;
-        entry.version_needed = file_info.version_needed;
-        entry.flags = file_info.flag;
-        entry.compressed_size = file_info.compressed_size;
-        entry.uncompressed_size = file_info.uncompressed_size;
+        entry.compression_method = static_cast<uint16_t>(file_info.compression_method);
+        entry.version_needed = static_cast<uint16_t>(file_info.version_needed);
+        entry.flags = static_cast<uint16_t>(file_info.flag);
+        entry.compressed_size = static_cast<uint32_t>(file_info.compressed_size);
+        entry.uncompressed_size = static_cast<uint32_t>(file_info.uncompressed_size);
         entry.is_encrypted = (file_info.flag & 0x01) != 0;
 
         // Unix file attribute constants
